+++ tests/test_query.py
@@ -1,5 +1,7 @@
 import pytest
 
+from textual.app import App, ComposeResult
+from textual.containers import Container
 from textual.widget import Widget
 from textual.css.query import InvalidQueryFormat, WrongType, NoMatches, TooManyMatches
 
@@ -50,17 +52,16 @@ def test_query():
         assert list(app.query(".frob")) == []
         assert list(app.query("#frob")) == []
 
-        assert app.query("App")
         assert not app.query("NotAnApp")
 
-        assert list(app.query("App")) == [app]
+        assert list(app.query("App")) == []  # Root is not included in queries
         assert list(app.query("#main")) == [main_view]
         assert list(app.query("View#main")) == [main_view]
         assert list(app.query("View2#help")) == [help_view]
         assert list(app.query("#widget1")) == [widget1]
-        assert list(app.query("#Widget1")) == [] # Note case.
+        assert list(app.query("#Widget1")) == []  # Note case.
         assert list(app.query("#widget2")) == [widget2]
-        assert list(app.query("#Widget2")) == [] # Note case.
+        assert list(app.query("#Widget2")) == []  # Note case.
 
         assert list(app.query("Widget.float")) == [sidebar, tooltip, helpbar]
         assert list(app.query(Widget).filter(".float")) == [sidebar, tooltip, helpbar]
@@ -110,10 +111,10 @@ def test_query():
             tooltip,
         ]
 
-        assert list(app.query("App,View")) == [app, main_view, sub_view, help_view]
+        assert list(app.query("View")) == [main_view, sub_view, help_view]
         assert list(app.query("#widget1, #widget2")) == [widget1, widget2]
         assert list(app.query("#widget1 , #widget2")) == [widget1, widget2]
-        assert list(app.query("#widget1, #widget2, App")) == [app, widget1, widget2]
+        assert list(app.query("#widget1, #widget2, App")) == [widget1, widget2]
 
         assert app.query(".float").first() == sidebar
         assert app.query(".float").last() == helpbar
@@ -130,14 +131,13 @@ def test_query():
 
 
 def test_query_classes():
-
     class App(Widget):
         pass
 
     class ClassTest(Widget):
         pass
 
-    CHILD_COUNT=100
+    CHILD_COUNT = 100
 
     # Create a fake app to hold everything else.
     app = App()
@@ -147,34 +147,35 @@ def test_query_classes():
         app._add_child(ClassTest(id=f"child{n}"))
 
     # Let's just be 100% sure everything was created fine.
-    assert len(app.query(ClassTest))==CHILD_COUNT
+    assert len(app.query(ClassTest)) == CHILD_COUNT
 
     # Now, let's check there are *no* children with the test class.
-    assert len(app.query(".test"))==0
+    assert len(app.query(".test")) == 0
 
     # Add the test class to everything and then check again.
     app.query(ClassTest).add_class("test")
-    assert len(app.query(".test"))==CHILD_COUNT
+    assert len(app.query(".test")) == CHILD_COUNT
 
     # Remove the test class from everything then try again.
     app.query(ClassTest).remove_class("test")
-    assert len(app.query(".test"))==0
+    assert len(app.query(".test")) == 0
 
     # Add the test class to everything using set_class.
     app.query(ClassTest).set_class(True, "test")
-    assert len(app.query(".test"))==CHILD_COUNT
+    assert len(app.query(".test")) == CHILD_COUNT
 
     # Remove the test class from everything using set_class.
     app.query(ClassTest).set_class(False, "test")
-    assert len(app.query(".test"))==0
+    assert len(app.query(".test")) == 0
 
     # Add the test class to everything using toggle_class.
     app.query(ClassTest).toggle_class("test")
-    assert len(app.query(".test"))==CHILD_COUNT
+    assert len(app.query(".test")) == CHILD_COUNT
 
     # Remove the test class from everything using toggle_class.
     app.query(ClassTest).toggle_class("test")
-    assert len(app.query(".test"))==0
+    assert len(app.query(".test")) == 0
+
 
 def test_invalid_query():
     class App(Widget):
@@ -187,3 +188,26 @@ def test_invalid_query():
 
     with pytest.raises(InvalidQueryFormat):
         app.query("#foo").exclude("#2")
+
+
+async def test_universal_selector_doesnt_select_self():
+    class ExampleApp(App):
+        def compose(self) -> ComposeResult:
+            yield Container(
+                Widget(
+                    Widget(),
+                    Widget(
+                        Widget(),
+                    ),
+                ),
+                Widget(),
+                id="root-container",
+            )
+
+    app = ExampleApp()
+    async with app.run_test():
+        container = app.query_one("#root-container", Container)
+        query = container.query("*")
+        results = list(query.results())
+        assert len(results) == 5
+        assert not any(node.id == "root-container" for node in results)